Updates to IMU data, quick library comment
Added data for the servos and the IMU with a rotational environment -
able to see the behaviour acting on it and quanitatively say what is
happening in the system. Tyaan in the previous commit has recalibrated
the control system and adapted the .csv format output on the serial
line.

Physically, the Teensy 3.2 has new breakouts so IMU boards and servos
can be replaced at the rocket launch site. Mainainability and robustness
has increased. Last things to do is to ensure the power source is
consistently on and to indicate that the system is about to launch -
i.e. LED after period of waiting time.