GPS Added Visualisations
Added map and visulations of the data. This was acheived using
www.gpsvisualizer.com. Signiifcant indication is the accuaracy of the
measurements in a 'flat' area and how they are more noisy at step hills
(around bottom left corner of plot). 600m range with significant
buildings and obstructions indicates a reliability in the range of the
module.